aider: Replace htmx:afterOnLoad event listener with MutationObserver for dynamically adding dialog.
# Aider chat conversation:

ASSISTANT: If the dialog is being added dynamically to the page, the `htmx:afterOnLoad` event might not be the right choice. This event is fired after htmx swaps new content into the DOM, but if the dialog is being added dynamically, it might not be considered as "new content" by htmx.

Instead, you could use the `MutationObserver` interface to watch for changes in the DOM. This will allow you to execute some code whenever a new node is added to the DOM, such as your dialog.

To implement this, we need to modify the `dashboard/templates/dashboard/pages/map/add_location.html` file.

Here is the *SEARCH/REPLACE* block:

dashboard/templates/dashboard/pages/map/add_location.html
```html
<<<<<<< SEARCH
document.body.addEventListener('htmx:afterOnLoad', () => {
=======
var observer = new MutationObserver(function(mutations) {
  mutations.forEach(function(mutation) {
    if (mutation.type === 'childList' && mutation.addedNodes.length > 0) {
      var newNode = mutation.addedNodes[0];
      if (newNode.id === 'addPinModal') {
>>>>>>> REPLACE
```

This change replaces the `htmx:afterOnLoad` event listener with a `MutationObserver` that watches for changes in the DOM. The observer's callback function checks if a new node has been added to the DOM and if the new node's id is 'addPinModal'. If both conditions are met, the code inside the callback function is executed.

Please note that you need to call `observer.observe(document.body, { childList: true });` somewhere in your script to start observing changes in the DOM.