aider: Create logout and signup templates for user authentication.
# Aider chat conversation:

USER: Create any other html templates that are needed, such as a signup or logout template. Only create them if django does not provide one by default.

ASSISTANT: Django does not provide default templates for login, logout, or signup. It only provides the views and forms that can be used in these templates.

We will need to create the following templates:

1. `registration/logout.html`: This template will be displayed after a user logs out. It can be a simple page that informs the user they have been logged out and provides a link to the login page.

2. `registration/signup.html`: This template will be used for user registration. It should include a form for the user to enter their username, email, and password.

Here are the *SEARCH/REPLACE* blocks to create these templates:

dashboard/templates/registration/logout.html
```html
<<<<<<< SEARCH
=======
{% extends "base_generic.html" %}

{% block content %}
  <h2>Logged out</h2>
  <p>You have been successfully logged out. Click <a href="{% url 'login' %}">here</a> to log in again.</p>
{% endblock %}
>>>>>>> REPLACE
```

dashboard/templates/registration/signup.html
```html
<<<<<<< SEARCH
=======
{% extends "base_generic.html" %}

{% block content %}
  <h2>Sign up</h2>
  <form method="post">
  {% csrf_token %}
    {{ form.as_p }}
    <button type="submit">Sign up</button>
  </form>
{% endblock %}
>>>>>>> REPLACE
```

Please note that these are very basic templates. You may want to customize them to match the look and feel of your site. Also, the signup template assumes that you have a signup view that passes an instance of Django's `UserCreationForm` (or a custom user registration form) to the template context as `form`. If this is not the case, you will need to adjust the template accordingly.