aider: Add an "upload" button near the user avatar to allow the user to upload an image as their avatar.
# Aider chat conversation:

USER: Create an "upload" button near the user avatar that allows the user to upload an image to be used as their avatar.

ASSISTANT: To add an "upload" button near the user avatar, we need to modify `dashboard/templates/dashboard/pages/profile/view.html`:

1. Add an "upload" button and a hidden file input field inside the avatar div. The button will trigger a click event on the hidden file input field when clicked, allowing the user to select an image file to upload.

Here is the *SEARCH/REPLACE* block:

dashboard/templates/dashboard/pages/profile/view.html
```html
<<<<<<< SEARCH
    <div class="avatar">
        {% if profile.avatar %}
            <img src="{{ profile.avatar.url }}" alt="Avatar" />
        {% else %}
            <div class="avatar-placeholder">
                {{ profile.user.first_name|first }}
            </div>
        {% endif %}
    </div>
=======
    <div class="avatar">
        {% if profile.avatar %}
            <img src="{{ profile.avatar.url }}" alt="Avatar" />
        {% else %}
            <div class="avatar-placeholder">
                {{ profile.user.first_name|first }}
            </div>
        {% endif %}
        <button id="upload-avatar-button" class="btn btn-primary">Upload</button>
        <input type="file" id="avatar-file-input" style="display: none;">
    </div>
>>>>>>> REPLACE
```

This will add an "upload" button and a hidden file input field inside the avatar div. When the "upload" button is clicked, it will trigger a click event on the hidden file input field, allowing the user to select an image file to upload.

Please note that you will need to add JavaScript to handle the file upload process and update the user's avatar. This includes listening for the "upload" button click event, triggering the file input click event, and sending a POST request to the server with the selected file.