aider: Send AJAX request to REST API when star is clicked and display notification upon response.
# Aider chat conversation:

ASSISTANT: To make this change we need to modify `dashboard/templates/dashboard/pages/map/index.html`:

1. Update the JavaScript code to send an AJAX request to the REST API when a star is clicked.
2. Display a notification to the user once the REST API returns a response.

Here is the *SEARCH/REPLACE* block:

dashboard/templates/dashboard/pages/map/index.html
```html
<<<<<<< SEARCH
<script>
    $(document).ready(function() {
        $('.stars input').on('change', function() {
            const star = $(this);
            const value = star.val();
            // TODO: Update the priority on the server
            console.log('New priority: ' + value);
        });
    });
</script>
=======
<script>
    $(document).ready(function() {
        $('.stars input').on('change', function() {
            const star = $(this);
            const value = star.val();
            const locationId = star.closest('.pin-popup').data('location-id');
            $.ajax({
                url: `/api/locations/${locationId}/`,
                type: 'PATCH',
                data: { priority: value },
                success: function() {
                    toastr.success('Priority updated successfully');
                },
                error: function() {
                    toastr.error('Failed to update priority');
                }
            });
        });
    });
</script>
>>>>>>> REPLACE
```

This change will send an AJAX request to the REST API when a star is clicked, updating the priority of the location. It will then display a success or error notification to the user depending on the response from the REST API. Note that this assumes the REST API endpoint for updating a location is `/api/locations/{locationId}/` and that the location ID is stored as a data attribute on the closest `.pin-popup` element. If this is not the case, you will need to adjust the code accordingly.