aider: Fixed accessing parent element's dataset property in pin popup field click event listener.
# Aider chat conversation:

USER: when clicking on a field value in the pin popup, I get this error: Uncaught TypeError: Cannot read properties of undefined (reading 'dataset')

ASSISTANT: The error message suggests that the script is trying to access the `dataset` property of an undefined object. This usually happens when you try to access a property of an object that doesn't exist.

In your case, it seems like the error is occurring in the event listener for the click event on the field value in the pin popup. The script is trying to access the `dataset` property of the parent element of the field, but it seems like the parent element is undefined.

To fix this issue, we need to ensure that the parent element exists before trying to access its `dataset` property.

Here is the *SEARCH/REPLACE* block to fix this issue:

dashboard/templates/dashboard/pages/map/index.html
```html
<<<<<<< SEARCH
                const field = this;
                const value = field.innerText;
                const input = document.createElement('input');
                input.value = value;
                input.name = field.parent.dataset.name;
=======
                const field = this;
                const value = field.innerText;
                const input = document.createElement('input');
                input.value = value;
                input.name = field.parentNode.dataset.name;
>>>>>>> REPLACE
```

In this block, I replaced `field.parent` with `field.parentNode` to correctly access the parent element of the field. This should fix the error you're seeing.