aider: Replace last visited date with human-readable format.
# Aider chat conversation:

ASSISTANT: To make this change we need to modify `dashboard/templates/dashboard/pages/map/index.html`:

1. Replace the line that displays the last visited date with a line that formats the date in a more human-readable format.

Here is the *SEARCH/REPLACE* block:

dashboard/templates/dashboard/pages/map/index.html
```html
<<<<<<< SEARCH
<div data-name="last_visited"><span class="label">Last Visited:</span> <span class="value">${pin.last_visited}</span></div>
=======
<div data-name="last_visited"><span class="label">Last Visited:</span> <span class="value">${new Date(pin.last_visited).toLocaleDateString()}</span></div>
>>>>>>> REPLACE
```

Please note that this change assumes that the `pin.last_visited` value is a string in a format that can be parsed by the JavaScript `Date` constructor. If this is not the case, additional changes may be necessary.